requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
gitdb2==2.0.3
GitPython==2.1.7
idna==2.6
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
oauthlib==2.0.4
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
smmap2==2.0.3
tweepy==3.5.0
urllib3==1.22
